Fix an issue in the previous change

+++ python/alm_wrapper.h
@@ -21,7 +21,7 @@ extern "C" {
                       const double lavec[3][3],
                       const double xcoord[][3],
                       const int kd[],
-                      size_t kind[]);
+                      int kind[]);
     void alm_set_verbosity(const int id,
                            const int verbosity);
     // void set_magnetic_params(const double* const * magmom,
