+++ src/alm.cpp
@@ -119,7 +119,7 @@ void ALM::set_periodicity(const int is_periodic[3]) const // PERIODIC
 void ALM::set_cell(const size_t nat,
                    const double lavec[3][3],
                    const double xcoord[][3],
-                   const size_t kind[],
+                   const int kind[],
                    const std::string kdname[]) const
 {
     system->set_supercell(lavec, nat, kind, xcoord);
